:truck: register 관련 type 공통 type 폴더로 이동
