+++ src/pages/index.ts
@@ -7,5 +7,5 @@ export { default as Login } from "./user/login/Login";
 export { default as VerifyCode } from "./user/findId/verification/VerifyCode";
 export { default as VerifyEmail } from "./user/findId/verification/VerifyEmail";
 export { default as ShowId } from "./user/findId/ShowId";
-export { default as validateId } from "./user/login/validation/id-validation";
-export { default as validatePw } from "./user/login/validation/pw-validation";
+export { default as validateId } from "../components/validation/id-validation";
+export { default as validatePw } from "../components/validation/pw-validation";
