+++ src/pages/user/index.ts
@@ -4,6 +4,6 @@ export { default as VerifyCode } from "./findId/verification/VerifyCode";
 export { default as VerifyEmail } from "./findId/verification/VerifyEmail";
 export { default as ShowId } from "./findId/ShowId";
 
-export { default as validateId } from "./login/validation/id-validation";
-export { default as validatePw } from "./login/validation/pw-validation";
+export { default as validateId } from "../../components/validation/id-validation";
+export { default as validatePw } from "../../components/validation/pw-validation";
 export { default as Register } from "./register/Regstier";
