+++ src/pages/user/login/Login.tsx
@@ -1,8 +1,8 @@
 import React, { useReducer, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 
-import validateId from "./validation/id-validation";
-import validatePw from "./validation/pw-validation";
+import validateId from "../../../components/validation/id-validation";
+import validatePw from "../../../components/validation/pw-validation";
 import UserWrapper from "../../../components/UserWrapper";
 
 const InitialLoginInfoStatus: LoginValidateInfo = {
