+++ src/pages/user/register/RegisterEmail.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import { useUserInfo } from "./context";
+import { useUserInfo } from "./register-context/context";
 import { Modal } from "../../../components";
 
 import { EMAIL_DOMAIN, InfoType } from "../../../types/enums";
