+++ src/pages/user/register/RegisterId.tsx
@@ -1,6 +1,6 @@
 import React, { useRef, useState } from "react";
 import { Modal } from "../../../components";
-import { useUserInfo } from "./context";
+import { useUserInfo } from "./register-context/context";
 
 enum InfoType {
   USER_NAME,
