+++ src/pages/user/register/RegisterName.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import { useUserInfo } from "./context";
+import { useUserInfo } from "./register-context/context";
 import { Modal } from "../../../components";
 
 enum InfoType {
